api-lib/pyproject.toml
[tool.poetry]
name = "api-lib"
version = "0.1.0"
description = ""
authors = ["Alexander Slessor <alexjslessor@gmail.com>"]
readme = "README.md"
packages = [
    { include = "api_lib" }
]
[tool.poetry.dependencies]
python = ">=3.12,<3.14"
fastapi = {version = "0.117.1", extras = ["standard"]}
pyjwt = "2.8.0"
aio-pika = "9.5.6"
pytz = "2024.1"
pydantic-settings = "2.9.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"